+++ src/components/cell-tooltip/CellTooltipText.js
@@ -10,13 +10,16 @@ export default function CellTooltipText(props) {
   } = props;
 
   const ref = useRef(null);
-  const offsetPercentage = 10;
+  const isNarrow = (parentWidth < 300);
   // Do collision detection based on the bounds of the tooltip ancestor element.
   useEffect(() => {
     const el = ref.current;
+    const offsetPercentage = isNarrow ? -5 : 10;
     const translateX = (x > parentWidth / 2) ? -(100 + offsetPercentage) : offsetPercentage;
     const translateY = (y > parentHeight / 2) ? -(100 + offsetPercentage) : offsetPercentage;
-    el.style.transform = `translateX(${translateX}%) translateY(${translateY}%)`;
+    const scale = isNarrow ? 0.75 : 1.0;
+    el.style.transform = `translateX(${translateX}%) translateY(${translateY}%) scale(${scale})`;
+    el.style.whiteSpace = (isNarrow ? 'normal' : 'nowrap');
   });
 
   return (